pyproject.toml
[project]
name = "word-wlof"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "google-genai>=1.28.0",
    "uvicorn>=0.35.0",
]
